rf: removed cast that can be handled implicit

+++ ProjectEulerCSharp/Problem010.cs
@@ -7,8 +7,8 @@ namespace ProjectEulerCSharp
     public class Problem010
     {
         [Theory]
-        [InlineData((long)10, (long)17)]
-        [InlineData((long)2000000, 142913828922)]
+        [InlineData(10, 17)]
+        [InlineData(2000000, 142913828922)]
         public void should_find_the_sum_of_all_the_primes(long upperBound, long expected)
         {
             1.To(upperBound)
